fix(store): handle missing parent directory in fs operations
Ensure that filesystem operations create parent directories when they do not exist, preventing failures when writing to nested paths that have not been explicitly created.

Auto-committed-on: dragonfly

+++ src/store/fs_ops.rs
@@ -935,7 +935,7 @@ impl DeepTraceStore for FsOps {
                 None => runs.clear(),
             }
         }
-        Ok(removed as u64)
+        Ok(removed)
     }
 }
